chore: Move main overview charts into their own component
